Refactor: python, cts: Move SimulStop into the _cts module.
No code changes.
